+++ CHANGELOG.md
@@ -13,6 +13,7 @@ adheres to [Semantic Versioning](https://semver.org/).
 
 ### :rocket: Added
 
+- Backport `tarfile` module
 - Support for CPython 3.9
 - Support for CPython 3.13 free-threaded